test: improve coverage for multi-provider LLM functions
